+++ spec/rubocop/cops/hash_literal_spec.rb
@@ -24,6 +24,16 @@ module Rubocop
                        ['test = Hash.new(3)'])
         expect(a.offences).to be_empty
       end
+
+      it 'does not crash when a method is called on super' do
+        inspect_source(a,
+                       'file.rb',
+                       ['class Derived < Base',
+                        '  def func',
+                        '    super.slice(1..2)',
+                        '  end',
+                        'end'])
+      end
     end
   end
 end