+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### Bugs fixed
 
+* [#90](https://github.com/bbatsov/rubocop/issues/90) Two cops crash when scanning code using super
+
 ### Misc
 
 * [#88](https://github.com/bbatsov/rubocop/issues/88) Abort gracefully when interrupted with Ctrl-C